requirements.txt
alembic==1.17.2
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.1.2
marshmallow-sqlalchemy==1.4.2
mysql-connector-python==9.5.0
packaging==25.0
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.2.1
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.4
